requirements.txt
requests
flask
mega.py